+++ CHANGELOG.md
@@ -5,9 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-## Added
-## Changed
-## Fixed
+### Added
+### Changed
+### Fixed
+
+## [1.1.1] - 2019-08-09
+### Fixed
+- Optimize common employee status view for more employees
+- Fix menu entry if no keyword and address is present
 
 ## [1.1.0] - 2019-08-03
 ### Added